requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
python-dotenv
bcrypt
python-jose
mysqlclient
alembic